delete unused code and fix
