+++ BALSAMIC/commands/report/base.py
@@ -2,8 +2,6 @@
 import click
 
 from BALSAMIC.commands.report.deliver import deliver as deliver_command
-from BALSAMIC.commands.report.status import status as status_command
-
 
 @click.group()
 @click.pass_context
@@ -11,6 +9,5 @@ def report(context: click.Context):
     """Command to generate delivery files and check analysis status."""
     pass
 
-
 report.add_command(deliver_command)
-report.add_command(status_command)
+
